refactor(#279): decompose tiling.rs into render/ submodules; split pane_ops; no behavior change
- tiling.rs: 303 -> 191 lines. Keeps layout/focus/zoom/drag/dispatch. Extracts
  pane-type render bodies to `src/render/{terminal_pane,app_pane,agent_pane}.rs`.
- render/terminal_pane.rs: exit-message handling, TerminalView setup, name
  bar + tab-dot overlay.
- render/app_pane.rs: AppRenderContext build + ProcessApp::ui dispatch.
- render/agent_pane.rs: thin wrapper over the shipped #288 `agent_pane::render_and_drain`.
- pane_ops.rs (~1059 lines) -> pane_ops/{mod,create,layout,workspace}.rs.
  mod.rs is 16 lines (module declarations); layout.rs owns splits/close/nav;
  create.rs owns pane creation + launch entry points; workspace.rs owns
  multi-context + on-disk workspace save.

Pure structural refactor. No behavior change. `impl PlexiApp { ... }` blocks
are split across files but all methods stay on the same type, so call sites
elsewhere in the crate are unchanged. Grey-square fix preserved: outer
`pane_ui` paints the pane background and shrinks into the inner UI; the
extracted renderers inherit that inner UI. Zoomed-pane placeholder preserved
as a painter-only fill.

`cargo build --release` clean, zero warnings. `cargo test` green (workspace
contains no Rust tests today; no test bodies modified). `just install-alpha`
succeeds.
